perf: implement lazy loading for dashboard components
- Extract heavy components into separate files (LiveTerminal, ExecutionStatus, ScanHistory)
- Use Next.js dynamic imports with SSR disabled for client-only components
- Add React.memo on ScanForm to prevent unnecessary re-renders
- Add Suspense with skeleton loaders for smooth loading experience
- Fix all ESLint errors (refs during render, setState in effect)
